+++ CSK_Module_Encoder/scripts/Sensors/Encoder/FlowConfig/Encoder_FlowConfig.lua
@@ -1,0 +1,18 @@
+-- Include all relevant FlowConfig scripts
+
+--*****************************************************************
+-- Here you will find all the required content to provide specific
+-- features of this module via the 'CSK FlowConfig'.
+--*****************************************************************
+
+require('Sensors.Encoder.FlowConfig.Encoder_OnNewData')
+
+--- Function to react if FlowConfig was updated
+local function handleOnClearOldFlow()
+  if _G.availableAPIs.default and _G.availableAPIs.specific then
+    if encoder_Model.parameters.flowConfigPriority then
+      CSK_Encoder.clearFlowConfigRelevantConfiguration()
+    end
+  end
+end
+Script.register('CSK_FlowConfig.OnClearOldFlow', handleOnClearOldFlow)