+++ CSK_Module_Encoder/scripts/Sensors/Encoder/helper/checkAPIs.lua
@@ -5,16 +5,16 @@
 
 local availableAPIs = {}
 
+-- Function to load all default APIs
 local function loadAPIs()
   CSK_Encoder = require 'API.CSK_Encoder'
 
-  Container = require 'API.Container'
-  DateTime = require 'API.DateTime'
-  Engine = require 'API.Engine'
-  Flow = require 'API.Flow'
   Log = require 'API.Log'
   Log.Handler = require 'API.Log.Handler'
   Log.SharedLogger = require 'API.Log.SharedLogger'
+
+  Container = require 'API.Container'
+  Engine = require 'API.Engine'
   Object = require 'API.Object'
   Timer = require 'API.Timer'
 
@@ -25,18 +25,23 @@ local function loadAPIs()
       CSK_PersistentData = require 'API.CSK_PersistentData'
     elseif appList[i] == 'CSK_Module_UserManagement' then
       CSK_UserManagement = require 'API.CSK_UserManagement'
+    elseif appList[i] == 'CSK_Module_FlowConfig' then
+      CSK_FlowConfig = require 'API.CSK_FlowConfig'
     end
   end
 end
 
+-- Function to load specific APIs
 local function loadSpecificAPIs()
   -- If you want to check for specific APIs/functions supported on the device the module is running, place relevant APIs here
-  -- e.g.:
+  DateTime = require 'API.DateTime'
+  Flow = require 'API.Flow'
   Conveyor = require 'API.Conveyor'
   Conveyor.Timeout = require 'API.Conveyor.Timeout'
   Encoder = require 'API.Encoder'
 end
 
+-- Function to load related APIs
 local function checkRelatedAPIs()
   CSK_DigitalIOManager = require 'API.CSK_DigitalIOManager'
 end